refactored classes to make better sense
