+++ app/fixtures/texttest_fixture.php
@@ -9,7 +9,6 @@ use GildedRose\Repository\ItemRepository;
 use GildedRose\Serializer\ItemNormalizer;
 use GildedRose\Serializer\ItemsNormalizer;
 use GildedRose\StockManager;
-use GildedRose\Updater\StockProcessor;
 use Symfony\Component\Serializer\Encoder\CsvEncoder;
 
 $itemRepository = new ItemRepository(
@@ -23,11 +22,21 @@ $filepath = __DIR__ . '/../data/testfixture.csv';
 $itemRepository->setItems($filepath, 'csv');
 $items = $itemRepository->getItems();
 
-$stockManager = new StockManager(new StockProcessor());
+$stockManager = new StockManager();
 
 $days = 2;
 if (count($argv) > 1) {
     $days = (int) $argv[1];
 }
 
-$stockManager->process($items, intval($days));
+echo 'OMGHAI!' . PHP_EOL;
+
+for ($i = 0; $i < $days; $i++) {
+    echo "-------- day ${i} --------" . PHP_EOL;
+    echo 'name, sellIn, quality' . PHP_EOL;
+    foreach ($items as $item) {
+        echo $item . PHP_EOL;
+    }
+    echo PHP_EOL;
+    $stockManager->updateAll($items);
+}
